+++ backend/models/imagesModel.js
@@ -18,15 +18,3 @@ const Image = sequelize.define('image', {
 Image.hasMany(RankDonate, { foreignKey: 'imageId' })
 
 module.exports = Image;
-
-// const Sequelize = require('sequelize');
-// const sequelize = require('../db');
-// const { DataTypes } = require('sequelize');
-
-// // Определение модели для сохранения фото в БД
-// const Photo = sequelize.define('photo', {
-//   filename: Sequelize.STRING,
-//   data: Sequelize.BLOB('long') // Бинарные данные для хранения изображения
-// });
-
-// module.exports = Photo;
