+++ backend/routers/adminRouter.js
@@ -1,12 +1,8 @@
 const Router = require ('express');
 const router = new Router();
-
 const roleMiddleware = require('../middleware/roleMiddleware');
-
 const adminController = require('../controllers/admin/adminController');
 const ranksController = require('../controllers/admin/ranksController');
-const imageController = require('../controllers/admin/imageController');
-
 const multer = require('multer');
 
 const storage = multer.memoryStorage();
@@ -26,15 +22,6 @@ router.get('/get-all-ranks', ranksController.getAllRanks);
 // http://localhost:5000/api/admin/get-one-rank?rankId=1ss
 router.get('/get-one-rank', roleMiddleware(["ADMIN"]), ranksController.getOneRank);
 
-
-
-// http://localhost:5000/api/admin/send-img/(imageName)
-// router.get('/send-img/:filename', imageController.sendImg);
-
-// router.post('/upload-img', imageController.)
-
-
-
 // http://localhost:5000/api/admin/add-rank
 router.post('/add-rank', upload.single('photo'), roleMiddleware(["ADMIN"]), ranksController.addRank);
 
@@ -42,6 +29,6 @@ router.post('/add-rank', upload.single('photo'), roleMiddleware(["ADMIN"]), rank
 router.delete('/delete-rank/:rankId', roleMiddleware(["ADMIN"]), ranksController.deleteRank);
 
 // http://localhost:5000/api/admin/update-rank
-router.put('/update-rank', ranksController.updateRank);
+router.put('/update-rank', roleMiddleware(["ADMIN"]), ranksController.updateRank);
 
 module.exports = router;
